Add fraction digit and refractor sales calculator
